Revert "first App on Kotlin"
