+++ app/build.gradle
@@ -1,12 +1,8 @@
 apply plugin: 'com.android.application'
+
 apply plugin: 'kotlin-android'
+
 apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-kapt'
-
-androidExtensions {
-    experimental = true
-
-}
 
 android {
     compileSdkVersion 29
@@ -29,32 +25,12 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.core:core-ktx:1.3.0-alpha01'
+    implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.core:core-ktx:1.0.2'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-    androidTestImplementation 'androidx.test:runner:1.3.0-alpha03'
-
+    androidTestImplementation 'androidx.test.ext:junit:1.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    implementation 'androidx.lifecycle:lifecycle-runtime:2.2.0'
-
-    kapt "android.arch.lifecycle:compiler:1.1.1"
-    implementation 'androidx.recyclerview:recyclerview:1.1.0'
-    implementation 'androidx.cardview:cardview:1.0.0'
-    //annotationProcessor 'androidx.annotation:annotation:1.1.0'
-    implementation 'androidx.coordinatorlayout:coordinatorlayout:1.1.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta4'
-
-
-//    implementation'com.githang:behavior-master:0.1'
-    // noinspection GradleCompatible
-    implementation "com.android.support:design:28.0.0"
-//    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-//    implementation "com.android.support:cardview-v7:28.0.0"
-//implementation 'com.android.support:coordinatorlayout:28.0.0'
-
-
 }
