--- app/src/main/java/com/example/kotlinlesson1/data/entity/Note.kt
@@ -1,34 +1,0 @@
-package com.example.kotlinlesson1.data.entity
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-import java.util.*
-
-@Parcelize
-data class Note(
-    val id: String,
-    val title: String = "",
-    val text: String = "",
-    val color: Color = Color.GREEN,
-    val lastChanged: Date = Date()
-) : Parcelable {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-        other as Note
-        if (id != other.id) return false
-        return true
-    }
-
-    enum class Color {
-        WHITE,
-        YELLOW,
-        GREEN,
-        BLUE,
-        RED,
-        VIOLET,
-        PINK
-
-    }
-    fun copy(){}
-}
