--- app/src/main/java/com/example/kotlinlesson1/ui/main/MainViewState.kt
@@ -1,5 +1,0 @@
-package com.example.kotlinlesson1.ui.main
-
-import com.example.kotlinlesson1.data.entity.Note
-
-data class MainViewState(val notes:List<Note>)
