+++ app/src/main/res/values/colors.xml
@@ -1,17 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#21928A</color>
+    <color name="colorPrimary">#008577</color>
     <color name="colorPrimaryDark">#00574B</color>
     <color name="colorAccent">#D81B60</color>
-    <color name="color_note">#D81B60</color>
-
-
-   <color name="WHITE">#FFFFFF</color>
-    <color name="YELLOW">#FFFF00</color>
-    <color name="GREEN">#00FF00</color>
-    <color name="BLUE">#0000FF</color>
-    <color name="RED">#FF0000</color>
-    <color name="VIOLET">#7B1FA2</color>
-    <color name="PINK">#FF00FF</color>
-
 </resources>
