--- app/src/main/res/values/dimens.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <dimen name="note_text">8sp</dimen>
-    <dimen name="margin_note">8dp</dimen>
-</resources>
