+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,3 @@
 <resources>
     <string name="app_name">KotlinLesson1</string>
-    <string name="title_app">notes</string>
-
-
 </resources>
